requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==5.2.0
